patient_api/requirements.txt
APScheduler==3.6.3
bcrypt==3.2.0
blinker==1.4
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.9
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.4
gcloud==0.18.3
googleapis-common-protos==1.52.0
gunicorn==19.9.0
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
PyJWT==1.7.1
MarkupSafe==1.1.1
marshmallow==3.9.1
nested-lookup==0.2.21
numpy==1.19.3
oauth2client==4.1.2
opencv-python==4.5.5.62
protobuf==3.14.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.9
Pyrebase4==4.4.1
pytz==2020.4
requests==2.25.0
requests-toolbelt==0.9.1
rsa==4.6
scipy==1.5.4
six==1.15.0
SQLAlchemy==1.3.20
tzlocal==2.1
urllib3==1.26.2
Werkzeug==1.0.1
